pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "allauthdropin"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.12"
keywords = []
authors = [
  { name = "Thomas Weholt", email = "thomas@weholt.org" },
]
classifiers = [  
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

dependencies = [
  "django",
  "django-crispy-forms", 
  "crispy-bootstrap5", 
  "django-allauth",
]

[project.optional-dependencies]
dev = [
  "hatch",
  "black",
  "isort",
]

[tool.hatch.build.targets.sdist]
include = [
  "src/allauthdropin",
]
